+++ src/main/java/io/github/xf8b/morefeatures/world/gen/MoreFeaturesGeneration.java
@@ -2,19 +2,23 @@ package io.github.xf8b.morefeatures.world.gen;
 
 import io.github.xf8b.morefeatures.MoreFeaturesRegistries;
 import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
+import io.github.xf8b.morefeatures.world.feature.LemonTree;
+import io.github.xf8b.morefeatures.world.feature.OrangeTree;
 import net.minecraft.world.biome.Biome;
+import net.minecraft.world.biome.Biomes;
 import net.minecraft.world.gen.GenerationStage;
 import net.minecraft.world.gen.feature.Feature;
 import net.minecraft.world.gen.feature.OreFeatureConfig;
+import net.minecraft.world.gen.placement.AtSurfaceWithExtraConfig;
 import net.minecraft.world.gen.placement.ConfiguredPlacement;
 import net.minecraft.world.gen.placement.CountRangeConfig;
 import net.minecraft.world.gen.placement.Placement;
 import net.minecraftforge.registries.ForgeRegistries;
 
-public class MoreFeaturesOreGeneration {
-    public static void generateOre() {
+public class MoreFeaturesGeneration {
+    public static void generateFeatures() {
         for (Biome biome : ForgeRegistries.BIOMES) {
-            if (MoreFeaturesConfig.isOreGenerationEnabled) {
+            if (MoreFeaturesConfig.isFeatureGenerationEnabled) {
                 ConfiguredPlacement sapphireOreConfig = Placement.COUNT_RANGE
                         .configure(new CountRangeConfig(1, 0, 0, 12));
                 biome.addFeature(GenerationStage.Decoration.UNDERGROUND_ORES, Feature.ORE
@@ -25,8 +29,19 @@ public class MoreFeaturesOreGeneration {
                 biome.addFeature(GenerationStage.Decoration.UNDERGROUND_ORES, Feature.ORE
                         .withConfiguration(new OreFeatureConfig(OreFeatureConfig.FillerBlockType.NATURAL_STONE, MoreFeaturesRegistries.ASBESTOS.get().getDefaultState(), 6))
                         .withPlacement(asbestosConfig));
+                if (biome == Biomes.PLAINS) {
+                    ConfiguredPlacement lemonTreeConfig = Placement.COUNT_EXTRA_HEIGHTMAP.
+                            configure(new AtSurfaceWithExtraConfig(0, 0.05F, 1));
+                    biome.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, Feature.NORMAL_TREE
+                            .withConfiguration(LemonTree.LEMON_TREE_CONFIG)
+                            .withPlacement(lemonTreeConfig));
+                    ConfiguredPlacement orangeTreeConfig = Placement.COUNT_EXTRA_HEIGHTMAP.
+                            configure(new AtSurfaceWithExtraConfig(0, 0.05F, 1));
+                    biome.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, Feature.NORMAL_TREE
+                            .withConfiguration(OrangeTree.ORANGE_TREE_CONFIG)
+                            .withPlacement(orangeTreeConfig));
+                }
             }
         }
     }
-
 }